[Modify] 각 파일에 있던 Interface를 types/types.ts 로 수정
