+++ components/Product/Detil/ProductDetailMainImage.tsx
@@ -1,9 +1,6 @@
 // 상품 상세 페이지에 해당 상품에 대한 메인 이미지 컴포넌트
 
 import Image from 'next/image';
-export interface PostPageTypes {
-  params: { id: string };
-}
 
 export default async function ProductDetailMainImage({ id }: { id: string }) {
   const res = await fetch(`${process.env.NEXT_PUBLIC_DEPLOY_URL}/api/product/${id}`, { cache: 'no-store' });
